+++ config/docs.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'basic_auth' => [
+        'username' => env('DOCS_BASIC_AUTH_USERNAME'),
+        'password' => env('DOCS_BASIC_AUTH_PASSWORD'),
+    ],
+];